Create Assignment3
Your program should compute the correct present value for the investment on the line, and then write a report for that investment to the report file. The report should be titled “Report on the investment of” then the individual's name. The next line should state the desired final value, the length of the investment, the interest rate, and then the present value of that investment computed by the previous program. The last part of the report should be a nicely formatted table that gives the increase in the value of the investment from year 1 until the end of the investment term.
Your program should process the entire file with information for several investors and save a report for each investor to the report file.