Create Assignment4
Part I:
Write a function drawSmiley(center, size, win) that draws a smiley face. The smiley face will be drawn in the GraphWin win with the center of the face at the Point center. The float variable size will determine the size of the face.
Part II:
Write a program that allows the user to draw a smiley face in a GraphWin by clicking on the window. The point clicked will be the center of the smiley face. Use your function drawSmiley.