Create Assignment5
Students are assigned to the Freshman, Sophomore, Junior, or Senior class according to the number of credits earned. Some universities use the credit levels in this table:
Number of credits>>>>>>>>>>>>Class
Between 0 and 27, inclusive>>>>>>Freshman
Between 28 and 55, inclusive>>>>>Sophomore
Between 56 and 87, inclusive>>>>>Junior
88 or more>>>>>>>>>>>>>>>>Senior
Write a program that gets as input from the user the name of a students and the number of credits that student has completed. You can assume that the number of credits is not negative. Then the program should print a sentence with the student’s name and the class the student belongs to. Be sure the output is a complete sentence, like “Luis Torres is a Sophomore.”